Updates for 2.4.1 release.

+++ RadeonGPUAnalyzerCLI/Src/kc_cli_commander_vulkan.cpp
@@ -963,7 +963,11 @@ bool KcCliCommanderVulkan::RunPostCompileSteps(const Config& config)
         }
     }
 
-    DeleteTempFiles();
+    if (!config.should_retain_temp_files)
+    {
+        DeleteTempFiles();
+    }
+
     return ret;
 }
